inmate: x86: fixup apic_freq to be unsigned long
apic_timer_init is expected to return unsigned long but was returning
unsigned long long. As the math should be safe with apic_freq being
unsigned long, the type is simply adjusted to match the function
declaration.

Fixes: 29b8b410c3e7 ("inmates: x86: Allow to bypass TSC and APIC timer
calibration")

+++ inmates/lib/x86/timing.c
@@ -104,7 +104,7 @@ void delay_us(unsigned long microsecs)
 
 unsigned long apic_timer_init(unsigned int vector)
 {
-	unsigned long long apic_freq;
+	unsigned long apic_freq;
 	unsigned long start, end;
 	unsigned long tmr;
